pyproject.toml
[tool.poetry]
name = "us-sale"
version = "0.1.0"
description = "US Sale Data Analytics"
authors = ["Narapady <chhuoynarapady@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
polar = "^0.0.118"
pandas = "2.1"
tabulate = "^0.9.0"
google-cloud-bigquery = "^3.13.0"
dbt-bigquery = "^1.7.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
dash = "^2.14.1"
plotly = "^5.18.0"
dash-bootstrap-components = "^1.5.0"
db-dtypes = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
